move get_plan_stream to spec_utils; dedup spec worker delegation into base
